In dotfiles:   * Update logout alias for GNOME 3.0 (Trac: #1027)
git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@25379 728af825-273c-0410-89f9-f7d3b574a069

+++ debathena/debathena/dotfiles/bashrc
@@ -106,7 +106,11 @@ fi
 # Set up standard bash shell initializations
 
 if [ "${XSESSION+set}" = set ]; then
+   if [ -x /usr/bin/gnome-session-quit ]; then
+	logout () { gnome-session-quit --no-prompt; exit; } # logout for X
+   else   
 	logout () { gnome-session-save --kill --silent; exit; } # logout for X
+   fi
 fi
 
 #   aliases dealing with adding locker programs
